split active and aggregate resources and fixed a bunch of bugs
